switch away from mtime

+++ lib/Event.ml
@@ -19,7 +19,7 @@
 module Hash = Digest.MD5
 
 type t =
-  | Point of (int64 * Stack_trace.raw_stack_trace)
+  | Point of (float * Stack_trace.raw_stack_trace)
   | Partial of {id: Hash.t; bytes: Bytes.t; part: int; part_count: int}
 
 type marshaled = bytes * int
